requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.8
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
